log errors inspect4py

+++ src/scc/commands/extract_metadata.py
@@ -22,6 +22,7 @@ def fetch(repos_csv, output, use_inspect4py):
 
     print("It may take a while... Depends on repository size and Github API limitations.")
     failed_repos = []
+    failed_repos_i4p = []
     for repo_url in progressbar(repos_url, redirect_stdout=True):
         try:
             print(f"Extracting metadata from {repo_url}")
@@ -33,22 +34,33 @@ def fetch(repos_csv, output, use_inspect4py):
             print(f"ERROR: Could not extract metadata from {repo_url}")
             failed_repos.append(repo_url)
         
-        if use_inspect4py and 'Python' in metadata["languages"]["excerpt"]:
-            subprocess.call(
-                                f'cd {output} && git clone {repo_url} && inspect4py -i {str(repo_url).split("/")[-1]} -o inspect4py_tmp -si && cd ..', 
-                                shell = True, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT
-                            )
-            with open(f'{output}/inspect4py_tmp/directory_info.json') as f:
-                ins4py = json.load(f)
-            shutil.rmtree(f'{output}/inspect4py_tmp', ignore_errors=False, onerror=None)
-            shutil.rmtree(f'{output}/{str(repo_url).split("/")[-1]}', ignore_errors=False, onerror=None)
-            metadata["inspect4py"] = {}
-            metadata["inspect4py"]["software_type"] = ins4py["software_type"]
+        if use_inspect4py and 'languages' in metadata and 'Python' in metadata["languages"]["excerpt"]:
+            try:
+                subprocess.call(
+                                    f'cd {output} && git clone {repo_url} && inspect4py -i {str(repo_url).split("/")[-1]} -o inspect4py_tmp -si && cd ..', 
+                                    shell = True, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT
+                                )
+                with open(f'{output}/inspect4py_tmp/directory_info.json') as f:
+                    ins4py = json.load(f)
+                shutil.rmtree(f'{output}/inspect4py_tmp', ignore_errors=False, onerror=None)
+                shutil.rmtree(f'{output}/{str(repo_url).split("/")[-1]}', ignore_errors=False, onerror=None)
+                metadata["inspect4py"] = {}
+                metadata["inspect4py"]["software_type"] = ins4py["software_type"]
+            except:
+                shutil.rmtree(f'{output}/{str(repo_url).split("/")[-1]}', ignore_errors=False, onerror=None)
+                print(f"ERROR: Could not run inspect4py for {repo_url}")
+                failed_repos_i4p.append(repo_url)
+
             
         repo_full_name = (repo_url[19:]).replace("/", "_").replace(".","-")
         with open(f"{output}/{repo_full_name}.json", 'w') as repo_metadata:
             json.dump(metadata, repo_metadata, indent = 4)
-    
+
+    if len(failed_repos_i4p) > 0:
+        print("ERROR: inspect4py could not be ran in the following repo/s:")
+        for fr in failed_repos_i4p:
+            print(fr)
+
     if len(failed_repos) > 0:
         print("ERROR: metadata could not be extracted from the following repo/s:")
         for fr in failed_repos:
